test(dev-tools): split remaining oversized orchestrator-state tests
Relocate the orchestrator-state payload-shape tests and their helpers
into a new sibling file so test_validate_orchestration_artifacts.py
returns to the 500-line file-size cap, with no change to test behavior
or coverage.

Refs: #275
