+++ docs/features/active/2026-07-02-epic-orchestrate-275/evidence/qa-gates/final-python-format.2026-07-03T00-31.md
@@ -1,0 +1,10 @@
+# Final Python Format Check (Remediation Cycle 2)
+
+Timestamp: 2026-07-03T00-31
+
+Command: `poetry run black --check scripts/dev_tools tests/scripts/dev_tools`
+
+EXIT_CODE: 0
+
+Output Summary: `211 files would be left unchanged.` Zero files require reformatting
+(211 = 210 baseline files + 1 new sibling test file added by [P1-T1]).
