+++ docs/features/active/2026-07-02-epic-orchestrate-275/evidence/qa-gates/final-python-lint.2026-07-03T00-32.md
@@ -1,0 +1,10 @@
+# Final Python Lint Check (Remediation Cycle 2)
+
+Timestamp: 2026-07-03T00-32
+
+Command: `poetry run ruff check scripts/dev_tools tests/scripts/dev_tools`
+
+EXIT_CODE: 0
+
+Output Summary: `All checks passed!` — 0 lint violations after the fix, including the new
+sibling test file `test_validate_orchestration_artifacts_state_shape.py`.
