+++ docs/features/active/2026-07-02-epic-orchestrate-275/evidence/qa-gates/final-python-test.2026-07-03T00-34.md
@@ -1,0 +1,15 @@
+# Final Python Test Run, with Coverage (Remediation Cycle 2)
+
+Timestamp: 2026-07-03T00-34
+
+Command: `poetry run pytest --cov=scripts.dev_tools --cov-branch --cov-report=term-missing tests/scripts/dev_tools`
+
+EXIT_CODE: 0
+
+Output Summary: **1192 passed, 19 skipped, 0 failed** — identical to the P0-T7 baseline
+(1192 passed, 19 skipped, 0 failed), confirming net-zero test-count change from this
+relocation-only fix.
+
+Coverage `TOTAL` row (combined statement + branch): `Stmts=9032, Miss=1242, Branch=3250,
+BrPart=447, Cover=83%` — identical to the P0-T7 baseline figure of `Cover=83%`. No coverage
+regression relative to baseline.
