+++ docs/features/active/2026-07-02-epic-orchestrate-275/evidence/qa-gates/final-python-typecheck.2026-07-03T00-33.md
@@ -1,0 +1,11 @@
+# Final Python Type-Check (Remediation Cycle 2)
+
+Timestamp: 2026-07-03T00-33
+
+Command: `poetry run pyright scripts/dev_tools tests/scripts/dev_tools`
+
+EXIT_CODE: 0
+
+Output Summary: `0 errors, 0 warnings, 0 informations`. A benign notice about a newer
+Pyright release being available (v1.1.409 -> v1.1.411) is printed but does not affect the
+exit code or the zero-error result.
