+++ docs/features/active/2026-07-02-epic-orchestrate-275/evidence/qa-gates/python-linecount-result.2026-07-03T00-30.md
@@ -1,0 +1,17 @@
+# Python Line-Count Result — Fix Target (Remediation Cycle 2, Post-Fix)
+
+Timestamp: 2026-07-03T00-30
+
+Command: `(Get-Content tests/scripts/dev_tools/test_validate_orchestration_artifacts.py | Measure-Object -Line).Lines`
+
+EXIT_CODE: 0
+
+Output Summary: The literal plan command returned `320`. As documented in
+`evidence/remediation-baseline/python-linecount-baseline.2026-07-03T00-21.md`,
+`Measure-Object -Line` undercounts blank lines. Cross-check via
+`(Get-Content ...).Count` and `wc -l` independently confirm the true post-fix line count is
+**381**. Both the literal command's value (320) and the cross-checked true value (381) are
+well under the 500-line cap.
+
+**Branch taken: Primary branch (target achieved).** The recorded value (381, true count) is
+≤ 500. Proceeding to [P1-T4].
