+++ docs/features/active/2026-07-02-epic-orchestrate-275/evidence/qa-gates/remediation-cycle-2026-07-03T00-15-final-summary.2026-07-03T00-40.md
@@ -1,0 +1,50 @@
+# Remediation Cycle 2 — Final Summary (epic-orchestrate #275)
+
+Timestamp: 2026-07-03T00-40
+
+## Outcome of the Single Fix
+
+**PASS.** `tests/scripts/dev_tools/test_validate_orchestration_artifacts.py` was reduced
+from 513 lines to 381 lines (true count; both confirmed via `wc -l` and
+`Get-Content .Count`), well under the 500-line cap, by relocating the 8 residual
+`orchestrator-state` payload-shape test functions and their 2 exclusive helper functions
+(`get_first_receipt`, `build_namespaced_orchestrator_state`) into a new sibling file,
+`tests/scripts/dev_tools/test_validate_orchestration_artifacts_state_shape.py` (154 lines).
+The Primary branch (target achieved) applied at [P1-T3]; the Escalation branch
+(file-size-exception-request) was not required. See
+`evidence/qa-gates/python-linecount-result.2026-07-03T00-30.md` and
+`evidence/qa-gates/python-test-split-cycle2.2026-07-03T00-35.md`.
+
+## Toolchain Pass/Fail State
+
+All four Python toolchain stages passed cleanly in a single pass with 0 errors/violations:
+
+- Format (Black): `EXIT_CODE 0` — 211 files unchanged.
+  (`evidence/qa-gates/final-python-format.2026-07-03T00-31.md`)
+- Lint (Ruff): `EXIT_CODE 0` — all checks passed.
+  (`evidence/qa-gates/final-python-lint.2026-07-03T00-32.md`)
+- Type-check (Pyright): `EXIT_CODE 0` — 0 errors, 0 warnings.
+  (`evidence/qa-gates/final-python-typecheck.2026-07-03T00-33.md`)
+- Test with coverage (Pytest): `EXIT_CODE 0` — 1192 passed, 19 skipped, 0 failed; combined
+  coverage 83%, identical to the P0-T7 baseline (no regression).
+  (`evidence/qa-gates/final-python-test.2026-07-03T00-34.md`)
+
+## Out-of-Scope Confirmation
+
+No PowerShell or TypeScript artifact was re-run in this cycle. Per
+`remediation-inputs.2026-07-03T00-15.md`, this cycle is a Python-only test-file
+reorganization; PowerShell/TypeScript coverage artifacts are unaffected and out of scope.
+
+## Checkbox State — Remains Unchecked Pending Re-Audit
+
+`spec.md` AC14 (all four quality toolchains pass with no coverage regression) and the
+Generic closing item "Toolchain pass completed (format -> lint -> type-check -> test)"
+**remain unchecked** (`- [ ]`) as of this cycle. Neither item was edited by this plan
+(see [P3-T1] and [P3-T2] confirmation tasks — `Select-String` returned exactly one match
+each, and `spec.md` itself was not modified).
+
+## Next Step
+
+The next step is a `feature-review` pass producing new `code-review`, `feature-audit`, and
+`policy-audit` artifacts at a new exit timestamp. That independent re-audit alone may
+re-evaluate and check off `spec.md`'s AC14 and the Generic "Toolchain pass completed" item.
