+++ docs/features/active/2026-07-02-epic-orchestrate-275/evidence/remediation-baseline/phase0-instructions-read.2026-07-03T00-20.md
@@ -1,0 +1,24 @@
+# Phase 0 — Policy Instructions Read (Remediation Cycle 2)
+
+Timestamp: 2026-07-03T00-20
+
+Policy Order:
+
+1. `CLAUDE.md`
+2. `.claude/rules/general-code-change.md`
+3. `.claude/rules/general-unit-test.md`
+4. `.claude/rules/python.md`
+5. `.claude/rules/python-suppressions.md`
+6. `.claude/rules/self-explanatory-code-commenting.md`
+
+Files read, in the order read:
+
+1. `CLAUDE.md`
+2. `.claude/rules/general-code-change.md`
+3. `.claude/rules/general-unit-test.md`
+4. `.claude/rules/python.md`
+5. `.claude/rules/python-suppressions.md`
+6. `.claude/rules/self-explanatory-code-commenting.md`
+
+Output Summary: All 6 policy files read successfully in the specified order prior to
+performing the [P1] fix in this remediation cycle.
