+++ docs/features/active/2026-07-02-epic-orchestrate-275/evidence/remediation-baseline/python-dispatch-linecount-baseline.2026-07-03T00-22.md
@@ -1,0 +1,16 @@
+# Python Line-Count Baseline — Existing Sibling Split File (Remediation Cycle 2)
+
+Timestamp: 2026-07-03T00-22
+
+Command: `(Get-Content tests/scripts/dev_tools/test_validate_orchestration_artifacts_dispatch.py | Measure-Object -Line).Lines`
+
+EXIT_CODE: 0
+
+Output Summary: The literal plan command returned `190`. As documented in
+`python-linecount-baseline.2026-07-03T00-21.md`, `Measure-Object -Line` undercounts blank
+lines. Cross-check via `wc -l` confirms the true line count of
+`tests/scripts/dev_tools/test_validate_orchestration_artifacts_dispatch.py` is **255**. This
+file is not modified by this cycle's fix ([P1-T1]/[P1-T2] only touch
+`test_validate_orchestration_artifacts.py` and add the new sibling
+`test_validate_orchestration_artifacts_state_shape.py`); its baseline is recorded here so
+[P1-T4] can confirm it remains unchanged after the fix.
