+++ docs/features/active/2026-07-02-epic-orchestrate-275/evidence/remediation-baseline/python-format-baseline.2026-07-03T00-23.md
@@ -1,0 +1,10 @@
+# Python Format Baseline (Remediation Cycle 2)
+
+Timestamp: 2026-07-03T00-23
+
+Command: `poetry run black --check scripts/dev_tools tests/scripts/dev_tools`
+
+EXIT_CODE: 0
+
+Output Summary: `210 files would be left unchanged.` No formatting violations detected in
+the baseline state.
