+++ docs/features/active/2026-07-02-epic-orchestrate-275/evidence/remediation-baseline/python-linecount-baseline.2026-07-03T00-21.md
@@ -1,0 +1,16 @@
+# Python Line-Count Baseline — Fix Target (Remediation Cycle 2)
+
+Timestamp: 2026-07-03T00-21
+
+Command: `(Get-Content tests/scripts/dev_tools/test_validate_orchestration_artifacts.py | Measure-Object -Line).Lines`
+
+EXIT_CODE: 0
+
+Output Summary: The literal plan command returned `409`. Cross-check confirms this is a
+known `Measure-Object -Line` undercount: the command counts newline characters within each
+line's string content and returns `0` for empty-string (blank) lines rather than `1`, so
+blank lines are silently excluded from the total. The file contains 104 blank lines (verified
+via `grep -c '^$'`) and 409 non-blank lines, for a true total of **513** lines — confirmed
+independently via `(Get-Content ... ).Count` (returns `513`) and `wc -l` (returns `513`).
+The baseline line count for this fix target is **513**, matching this task's stated
+acceptance value and the plan's stated Scope figure.
