+++ docs/features/active/2026-07-02-epic-orchestrate-275/evidence/remediation-baseline/python-lint-baseline.2026-07-03T00-24.md
@@ -1,0 +1,9 @@
+# Python Lint Baseline (Remediation Cycle 2)
+
+Timestamp: 2026-07-03T00-24
+
+Command: `poetry run ruff check scripts/dev_tools tests/scripts/dev_tools`
+
+EXIT_CODE: 0
+
+Output Summary: `All checks passed!` — 0 lint violations in the baseline state.
